Add baseline model and its feature extractor
In order to establish a baseline that tagging models can be
compared against, a naive model is implemented. This model
counts the number of times a word has been a tag and uses
the most frequent tag as the prediction. If a word has not
been before, it assigns it a noun tag.