Calculate accuracy of ambiguous and unknown words
Overall accuracy may be misleading and the real challenge is to
tag words that may have more than one meaning or have not been seen.

The ambigouous accuracy metric filters for words that have more than
one tag and measures a model accuracy for them.

The unknown accuracy metrics filters for words that were not seen
in the training data and measures the model accuracy with them.