Use best_path_simple in hmm.NLTKModel because it predicts better
Not sure why this is but using `best_path_simple` instead of
`best_path` returns better predictions by a great margin,
around 35% better. I would need to dive deeper into the library
and see what the actual differences are but this could be a bug
in the `nltk` library.

Updated the results as well.